pakolás már majdnem megvan
eljutottam addig hogy az állomásokon lévő vagonokra felrakom azt ami ott van ha nem az a végállomás, de ez nem igazán működik még, de így nehéz lenne szerintem megcsinálni, úgyhogy heurisztikus kereséssel kéne menni az állapotok között.

+++ nagybeadando/src/Train.cpp
@@ -26,7 +26,7 @@ Train::Train(std::string filename)
 
 bool Train::csatol(Wagon* w)
 {
-    if(_wagons.size()<(size_t)_maxWagons)
+    if(_wagons.size()<_maxWagons)
     {
         _wagons.push_back(w);
         return true;
